add three version files
